python/veritrade/functions/verify/requirements.txt
google-cloud-vision==0.35.0